+++ HISTORY.rst
@@ -10,6 +10,9 @@
 - Add test for GitLab issue #702
 - Don't set SourceColumns to the attribute name when no help text
   is defined.
+- Make it clear when a setting is being overrided by an environment
+  variable. Add better text explanation. Change the column to a check
+  mark widget.
 
 3.3.3 (2019-12-05)
 ==================